pivotal-questions/_archive/workshop-collab-tool/requirements.txt
# Workshop Collaborative Space Tool
# Install: pip install -r requirements.txt

# CLI
click>=8.0

# Config & validation
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0

# HTTP clients
httpx>=0.25

# Data processing
pandas>=2.0

# Coda API (also used by coda_org_unjournal)
codaio>=0.7
requests>=2.28

# Google Docs
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
